refactor: 기존 카드 제거
